src/main/java/com/yjr/ordering_system/service/impl/CategoryServiceImpl.java
package com.yjr.ordering_system.service.impl;

import com.yjr.ordering_system.entity.ProductCategory;
import com.yjr.ordering_system.repository.ProductCategoryRepository;
import com.yjr.ordering_system.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.ArrayList;
import java.util.List;

/**
 * 类目
 * @author yjr
 * @since 2020/10/28 16:19
 */
@Service
public class CategoryServiceImpl implements CategoryService {

    @Autowired
    ProductCategoryRepository repository;

    @Override
    public ProductCategory findOne(Integer id) {
        return repository.findOne(id);
    }

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        if (categoryTypeList.isEmpty()) {
            return new ArrayList<>();
        }
        return repository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return repository.save(productCategory);
    }
}
